// src/app/models/publicfigure.ts
export interface publicFigure {
    id: number;
    name?: string;
    datePublish?: string;
    area?: string;
    closingDays?: string;
    longDesc?: string;
    shortDesc?: string;
    coverImage?: string;
    cardThumbnail?: string;
    isCover?: boolean;
    likes?: number;
    dislikes?: number;
}